app/source/requirements.txt
flask
pymongo
pyyaml
